% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{calc_diagnostics}
\alias{calc_diagnostics}
\title{see whether MCMC chains have converged; if so, calculate effective sample size and burn-in}
\usage{
calc_diagnostics(filenames, check_freq, fixed, skip = 0)
}
\arguments{
\item{filenames}{character vector of filenames, each corresponding to one csv
file outputted by run_MCMC}

\item{check_freq:}{check whether convergence has occurred every check_freq iterations
along the chain}

\item{fixed:}{logical vector: vector whose entries are TRUE if the corresponding
parameter is fixed, FALSE if the parameter is fitted. i.e. parTab$fixed}

\item{skip:}{numeric vector either of length 1 or of same length as filenames:
skip this many entries from the start of each csv file}
}
\value{
if convergence has occurred, return a list with the elements
converged: logical = TRUE
burn_in: the number of burn-in iterations
combined_size: effective sample size combined across chains
if convergence has not occurred, return a list with the element
converged: logical = FALSE
max.prsf:maximum value of potential scale reduction factor across parameters
}
\description{
see whether MCMC chains have converged; if so, calculate effective sample size and burn-in
}
