% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{calc_DIC}
\alias{calc_DIC}
\title{DIC calculation}
\usage{
calc_DIC(lik.fun, chain)
}
\arguments{
\item{lik.fun}{the posterior calculating function}

\item{chain}{the MCMC chain to be tested}
}
\value{
a single DIC value
}
\description{
Given an MCMC chain, calculates the DIC
}

