% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{ess_diagnostics}
\alias{ess_diagnostics}
\title{Check ESS}
\usage{
ess_diagnostics(chain, threshold = 200)
}
\arguments{
\item{chain}{the MCMC chain to be tested}

\item{threshold}{the minimum allowable ESS}
}
\value{
a list containing the ESS sizes and flagging which parameters have ESS below the threshold
}
\description{
Checks the MCMC chain for effective sample sizes used the coda package, and highlights where these are less than a certain threshold
}

