% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{estimate_mode}
\alias{estimate_mode}
\title{Calculate mode}
\usage{
estimate_mode(x)
}
\arguments{
\item{x}{the vector to be tested}
}
\value{
the point estimate for the mode
}
\description{
Calculates the posterior mode based on density
}
