% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{gelman_diagnostics}
\alias{gelman_diagnostics}
\title{Check gelman diagnostics}
\usage{
gelman_diagnostics(chain, threshold = 1.15)
}
\arguments{
\item{chain}{the list of MCMC chains}

\item{threshold}{the threshold for the gelman diagnostic above which chains should be rerun}
}
\value{
a list of gelman diagnostics and highlighting worst parameters
}
\description{
Checks the gelman diagnostics for the given MCMC chain list. Also finds the parameter with the highest PSRF.
}

