% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{generate_start_tab}
\alias{generate_start_tab}
\title{Generate starting parameter table}
\usage{
generate_start_tab(par_tab)
}
\arguments{
\item{par_tab}{See \code{\link{example_par_tab}}}
}
\value{
a data frame matching par_tab
}
\description{
Generates a version of \code{par_tab} with random values between \code{lower_start} and \code{upper_start}
}
\examples{
data(example_par_tab)
start_tab <- generate_start_tab(example_par_tab)
}
\concept{mcmc}
