% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{get_best_pars}
\alias{get_best_pars}
\title{Get maximum likelihood parameters}
\usage{
get_best_pars(chain)
}
\arguments{
\item{chain}{the MCMC chain as saved by run_MCMC}
}
\value{
a named vector of model parameters
}
\description{
From an MCMC chain produced by run_MCMC, find the row with the highest log likelihood. Return the parameters (excluding the first and last columns, which are sampno and lnlike respectively) that give the highest likelihood.
}
