% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{get_index_pars}
\alias{get_index_pars}
\title{Get parameters based on index}
\usage{
get_index_pars(chain, index)
}
\arguments{
\item{chain}{the MCMC chain as saved by \code{\link{run_MCMC}}}

\item{index}{value for sampno, matching the sampno column from the MCMC chain, to be returned}
}
\value{
a named vector of model parameters
}
\description{
From an MCMC chain produced by \code{\link{run_MCMC}}, return the parameters from the specified sample number (excluding the first and last columns, which are sampno and lnlike respectively).
}
