% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{load_mcmc_chains}
\alias{load_mcmc_chains}
\title{Read in MCMC chains}
\usage{
load_mcmc_chains(
  location = "",
  parTab,
  unfixed = TRUE,
  thin = 1,
  burnin = 1e+05,
  multi = TRUE,
  chainNo = FALSE,
  PTchain = FALSE
)
}
\arguments{
\item{location}{Either the full file path to the MCMC chain containing directory, or a vector of file paths with the MCMC chainsa}

\item{parTab}{the parameter table that was used to solve the model (mainly used to find which were free parameters)}

\item{unfixed}{Boolean, if TRUE, only returns free parameters}

\item{thin}{thin chain by this much}

\item{burnin}{number of iterations to discard}

\item{multi}{if TRUE, looks for chains generated using the multivariate sampler. Otherwise, looks for univariate sampled chains.}

\item{chainNo}{if TRUE, adds the chain number to the MCMC chain as a column}

\item{PTchain}{DEV - if TRUE, looks for chains generated by the parallel tempering algorithm}
}
\value{
a list containing both the MCMC chains appended to each other, and an MCMC list.
}
\description{
Loads all available MCMC chains from the chosen working directory, allowing the user to specify properties of the return chain
}
