% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mcmc_param_check}
\alias{mcmc_param_check}
\title{MCMC chain parameter check}
\usage{
mcmc_param_check(mcmcPars, mvrPars)
}
\arguments{
\item{mcmcPars}{the vector of MCMC pars as expected by \link{\code{run_MCMC}}}

\item{mvrPars}{the list of multivarate MCMC pars as expected by \link{\code{run_MCMC}}}
}
\value{
a list of errors
}
\description{
Checks that the given settings for the MCMC chain are correct
}
