% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{median.quantile}
\alias{median.quantile}
\title{Find median}
\usage{
\method{median}{quantile}(x)
}
\arguments{
\item{x}{the vector to be analysed}
}
\value{
the quantiles
}
\description{
Finds the median and 95% quantiles for a vector
}

