% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ lazymcmc }
\alias{ lazymcmc }
\alias{ lazymcmc }
\title{ MCMC for the lazy }
\description{
A package for very generic MCMC in R. Using either univariate uniform or multivariate gaussian proposals, explores the posterior distribution for an arbritary set of model parameters specified in R. The focus is to allow an intuitive, flexible interface rather than fancy on fancy optimisations and syntax.
}
