% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{param_table_check}
\alias{param_table_check}
\title{Model parameter table check}
\usage{
param_table_check(parTab)
}
\arguments{
\item{parTab}{the parameter table controlling information such as bounds, initial values etc}
}
\value{
a list of strings with the errors of the parameter table
}
\description{
Takes the table of paramters used for model calculation and MCMC fitting, and returns a string error if any of the inputs are incorrectly formatted
}

