% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{parLapply_wrapper}
\alias{parLapply_wrapper}
\title{wrapper for parLapply for cluster}
\usage{
parLapply_wrapper(run_parallel, x, fun, ...)
}
\arguments{
\item{x}{first argument of lapply}

\item{fun}{second argument of lapply}

\item{run_parallel:}{logical: if TRUE, use parLapply, else use lapply}
}
\value{
output arguments of lapply
}
\description{
wrapper for parLapply for cluster
}
