% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{plot_single_chain}
\alias{plot_single_chain}
\title{Density and iteration plot single chain}
\usage{
plot_single_chain(chain, realPars = NULL)
}
\arguments{
\item{chain}{the single MCMC chain}

\item{realPars}{OPTIONAL - vector of real parameter values to add to plot}
}
\value{
a ggplot2 object
}
\description{
Plots the posterior density and MCMC trace plots for the given chain. Requires ggplot2,cowplot and reshape2
}
