% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{print_prctiles}
\alias{print_prctiles}
\title{calculate percentiles for fitted parameters}
\usage{
print_prctiles(chain, prctiles = c(0.025, 0.5, 0.975), par_names_plot)
}
\arguments{
\item{chain}{the MCMC chain: a data frame with n columns, where n is the number
of fitted parameters}

\item{prctiles}{a numeric vector of length m containing the percentiles to be calculated
(between 0 and 1)}

\item{par_names_plot}{parameter names for the output data frame}
}
\value{
a data frame with n rows and m columns containing the percentiles
}
\description{
calculate percentiles for fitted parameters
}
