% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{profile_likelihood_all}
\alias{profile_likelihood_all}
\title{calculate and plot profile likelihood around a point in parameter space}
\usage{
profile_likelihood_all(stuck_values, f, fixed, n_values, span,
  par_names_plot)
}
\arguments{
\item{stuck_values}{a named numeric vector of parameter values (length m)
around which to construct the profile likelihood}

\item{f}{a function handle to calculate the log likelihood. 
f <- CREATE_POSTERIOR_FUNC(parTab, data, PRIOR_FUNC)}

\item{fixed}{a numeric vector of length m indicating whether a parameter is 
fixed (1) or fitted (0)}

\item{n_values}{a numeric vector of length 1 indicating the number of points 
to scan over in 1D parameter space for each profile. Must take an integer value.}

\item{span}{a numeric vector of length 1 indicating the span over which to
scan in 1D parameter space for each profile (so we go span/2 on each side of
the central value)}

\item{par_names_plot}{a character vector of length m to label the horizontal
axes of the profile plots}
}
\value{
a list of ggplot objects containing the profile likelihood plots
}
\description{
calculate and plot profile likelihood around a point in parameter space
}
