% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{summarise_chain}
\alias{summarise_chain}
\title{Summarise MCMC chain statistics}
\usage{
summarise_chain(chain)
}
\arguments{
\item{chain}{the MCMC chain}
}
\value{
a data frame of MCMC statistics
}
\description{
Returns code statistics and quantiles
}
