% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{try_fread}
\alias{try_fread}
\title{data.table::fread with tryCatch}
\usage{
try_fread(filename, ...)
}
\arguments{
\item{filename}{character vector of length 1 ending in .csv}
}
\value{
the read data frame
}
\description{
data.table::fread sometimes crashes.  If crashing, try read.csv then converting.
}
