% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{univ_proposal_normal}
\alias{univ_proposal_normal}
\title{MCMC proposal function - univariate normal}
\usage{
univ_proposal_normal(values, sds, index)
}
\arguments{
\item{values}{a vector of the parameters to be explored}

\item{sds}{a vector of standard deviations of the proposal function}

\item{index}{numeric value for the index of the parameter to be moved from the param table and vector}
}
\value{
the parameter vector after step
}
\description{
Proposal function for MCMC random walk, drawing from normal distribution centered on the current value
}
