% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{write.list}
\alias{write.list}
\title{writes a list to a text file}
\usage{
write.list(list.to.write, filename, overwrite)
}
\arguments{
\item{list.to.write:}{list to write to file}

\item{filename:}{character vector of length 1: filename to write to}
}
\description{
writes a list to a text file
}
\details{
doesn't work for lists of lists
}
