% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{write_list}
\alias{write_list}
\title{writes a list to a text file}
\usage{
write_list(list_to_write, filename, overwrite)
}
\arguments{
\item{list_to_write:}{list to write to file}

\item{filename:}{character vector of length 1: filename to write to}
}
\description{
writes a list to a text file
}
\details{
doesn't work for lists of lists
}
